requirements.txt
django~=5.1.6
pyotp~=2.9.0
cryptography~=44.0.1
pyHanko~=0.26.0
pyhanko-certvalidator~=0.26.8
minio~=7.2.15
Flask~=3.1.0
Werkzeug~=3.1.3
SQLAlchemy~=2.0.38
config~=0.5.1
python-jose~=3.5.0
passlib~=1.7.4
fastapi~=0.115.12
requests~=2.32.3
uvicorn~=0.34.2
pydantic~=2.11.3
pywebview~=5.4
django-allauth~=65.9.0
PyJWT~=2.10.1